Fix yarn build, refactor for better readable files
